requirements.txt
ccxt==4.3.91
python-telegram-bot==21.4
python-dotenv==1.0.1
aiohttp==3.10.5
